examples moved to cgru root
git-svn-id: https://cgru.svn.sourceforge.net/svnroot/cgru@1042 0f86f178-d61e-0410-9d12-f6255c3815b8

+++ afanasy/trunk/src/project.qmake/build.sh
@@ -41,4 +41,5 @@ $QMAKE
 make
 
 # Copy python library:
-cp -fv libpyaf/libpyaf.so ../../bin/pyaf.so
+pyaf=libpyaf/libpyaf.so
+[ -f $pyaf ] && cp -fv libpyaf/libpyaf.so ../../bin/pyaf.so
